Add assert tracking to testharness.js
This keeps a global list of all the asserts that are triggered along
with details about the argument types and so on. We could make this
per test, but the idea is that keeping the list globally provides
additional informataion about ordering that might be relevant in some
cases.

To keep the arguments we are currently keeping either JSON.stringify
of the argument value or toString of the value. This isn't terribly
efficient, but the concern is that keeping the actual object alive
will increase memory usage and may affect tests.